Refactors creating new components
